feat: improvements to file upload
* Remove multi file upload stub as it can be implemented when the needed
* Store files in a separate directory for each user
* For each uploaded file a FileDescriptor DB entry is added
   Usage example: An activity is imported from a FIT file and the
   DB entry can reference this file. Or an image file is added to an
   activity.

To keep the API free of side effects, the upload API now returns a
file descriptor with a filename and the file id on new uploads. The API
user can later use this information with another API call to import
this file to, for example, a new activity. The upload API does not
automatically trigger an import of a FIT file, this must be done
client side.
This increases the testability for each API endpoint.